// src/types/token.ts
export interface TokenSecurity {
    mintAuthBurned: boolean;
    freezable: boolean;
    permanentDelegate: boolean;
    top10HoldersPercent: number;
}